refactor(gallery): move get images function to the db service
move get images function to the db service

MODULE-2-PART-4-mongodb
